+++ src/main/java/net/mrscauthd/beyond_earth/common/entities/IceSpitEntity.java
@@ -75,7 +75,7 @@ public class IceSpitEntity extends AbstractArrow implements ItemSupplier {
         double d1 = target.getX() - entity.getX();
         double d3 = target.getZ() - entity.getZ();
 
-        entityArrow.shoot(d1, d0 - entityArrow.getY() + (double) Math.sqrt(d1 * d1 + d3 * d3) * 0.2F, d3, 0.7f * 2, 12.0F);
+        entityArrow.shoot(d1, d0 - entityArrow.getY() + Math.sqrt(d1 * d1 + d3 * d3) * 0.2F, d3, 0.7f * 2, 12.0F);
         entityArrow.setSilent(true);
         entityArrow.setBaseDamage(damage);
         entityArrow.setKnockback(1);
